Refactor webhook routes and enhance pull request notification formatting
